refactor: Character Servis component unit test and index.test unit test
