+++ tests/components/CharacterService.test.ts
@@ -3,7 +3,7 @@ import { server } from '../mocks/server';
 import '@testing-library/jest-dom/extend-expect';
 import '@testing-library/jest-dom';
 import 'jest';
-import RickAndMortyService from '../../src/components/RickAndMortyService';
+import CharacterService from '../../src/components/CharacterService';
 import { persons as personData } from '../../src/data/data';
 
 let container: HTMLDivElement | null = null;
@@ -26,13 +26,13 @@ afterEach(() => {
 
 describe('RickAndMortyService Service', () => {
   it('handlePersonsFetched with provided wrong search value return default 4 data cards', async () => {
-    const rickAndMortyService = new RickAndMortyService();
+    const rickAndMortyService = new CharacterService();
     const persons = await rickAndMortyService.getPersons('qwert');
     expect(persons).toEqual(personData);
   });
 
   it('handlePersonsFetched with provided valid search value return not empty value', async () => {
-    const rickAndMortyService = new RickAndMortyService();
+    const rickAndMortyService = new CharacterService();
     const persons = await rickAndMortyService.getPersons('kyle');
     expect(persons).toHaveLength(1);
   });
