+++ tests/index.test.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import { unmountComponentAtNode } from 'react-dom';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
+import 'jest';
+import App from '../src/App';
+
+let container: HTMLDivElement | null = null;
+
+beforeEach(() => {
+  container = document.createElement('div') as HTMLDivElement;
+  document.body.appendChild(container);
+});
+
+afterEach(() => {
+  unmountComponentAtNode(container as Element);
+  if (container) {
+    unmountComponentAtNode(container);
+    container.remove();
+    container = null;
+  }
+});
+
+test('renders without crashing', () => {
+  render(
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  );
+
+  const card = screen.getAllByTestId('card');
+  expect(card.length).toBeGreaterThanOrEqual(4);
+});
